Add some stuff on function restriction
OK so I've hit a huge obstacle in trying to define a function that
checks if its input is equal to some number, since I think we need to
show that equality of mynat is decidable?  And idk how to do that.

That seems like an important thing to do though, and having done that,
I'm pretty sure my outline should work, given a fairly large investment
of actually writing and fleshing things out in Lean.