Add basic wlog_le theorem
This is like a little prototype. It doesn't even save space, since it's
such a pain to prove the proposition is symmetric, and you have to
re-introduce and assume everything in both cases.

However I do think it makes the proofs nicer, since I'd rather reduce
the conceptual duplication, and I like the idea of separately proving
symmetry.