Add tactic.lean to principia and top-level tactic dir
I made the tactic dir to learn how to make tactics, that is where
I have my propositional logic tactics.
Progress towards a wlog_le tactic is in principia/tactic.lean

I've annotated all the tactics I made, in the hope that they might
be possible to understand. Often I copied things that I didn't
understand so I'm not exactly sure what a metavariable is or
what the difference between one two or three backticks is