Some more work on natural subtraction
Lots of rephrasing in terms of lt and le.

In some ways this opens up some nice intuitive avenues for proofs
(eg sub_succ_rearrange is great for my intuition) but I think it does
increase the risk of overcomplicating certain things. Some proofs seem
to be too much of a case-bash.